Add Roslyn integration
Create a new Roslyn command and add it as `-roslyn`.
It uses the Roslyn C# interpreter to provide a REPL similar to the
already existing C# and Python REPLs. Once mature enough, it will replace
the C# REPL, which is currently being provided by Mono.CSharp.

The Roslyn REPL also has access to the database, IRC interface and
the configuration manager, so several access modifiers on existing
classes had to be changed.