Allow the bot to use multiple databases
This restructures the bot so that IRC Clients are now assigned a
StatsDatabaseManager instance. All message- and commandhandlers
will now use the Message.Client.StatsDatabase property to perform
database operations relevant to that message, instead of relying
on a DataFunctionSet object passed to it at initialisation time.

As a result, the DataFunctionSet class has been renamed to
StatsDatabaseManager to reflect its new role.